update CartItem, CartService,  CartController, OrderService, OneTimeScheduleService and OneTimeScheduleController

+++ src/main/java/com/project/petmanagement/petmanagement/services/OrderService.java
@@ -34,17 +34,26 @@ public class OrderService {
     }
 
     @Transactional(rollbackFor = {Exception.class})
-    public Order createOrder(OrderRequest orderRequest) throws Exception {
+    public Order createOrder(OrderRequest orderRequest) throws DataNotFoundException {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         JWTUserDetail jwtUserDetail = (JWTUserDetail) authentication.getPrincipal();
         Cart cart = cartRepository.findByUser(jwtUserDetail.getUser());
-        if (cart == null || cart.getCartItems().isEmpty()) {
+        boolean existedSelectedItem = false;
+        for (CartItem item : cart.getCartItems()) {
+            if (item.getSelected()) {
+                existedSelectedItem = true;
+                break;
+            }
+        }
+        if (cart.getCartItems().isEmpty()) {
             throw new DataNotFoundException("Cart is empty");
+        }
+        if (!existedSelectedItem) {
+            throw new DataNotFoundException("No item to create order");
         }
         Order order = Order.builder()
                 .user(cart.getUser())
                 .orderDate(new Date())
-                .totalPrice(cart.getTotalPrice())
                 .phone(orderRequest.getPhone())
                 .shippingAddress(orderRequest.getShippingAddress())
                 .paymentMethod(orderRequest.getPaymentMethod())
@@ -52,17 +61,22 @@ public class OrderService {
                 .build();
         order = orderRepository.save(order);
         List<CartItem> cartItems = cart.getCartItems();
+        double totalPrice = 0;
         for (CartItem cartItem : cartItems) {
-            OrderDetail orderDetail = OrderDetail.builder()
-                    .order(order)
-                    .product(cartItem.getProduct())
-                    .quantity(cartItem.getQuantity())
-                    .price(cartItem.getProduct().getPrice())
-                    .build();
-            orderDetailRepository.save(orderDetail);
-            cartItemRepository.deleteById(cartItem.getId());
+            if (cartItem.getSelected()) {
+                OrderDetail orderDetail = OrderDetail.builder()
+                        .order(order)
+                        .product(cartItem.getProduct())
+                        .quantity(cartItem.getQuantity())
+                        .price(cartItem.getProduct().getPrice())
+                        .build();
+                orderDetailRepository.save(orderDetail);
+                cartItemRepository.deleteById(cartItem.getId());
+                totalPrice += orderDetail.getPrice() * orderDetail.getQuantity();
+            }
         }
-        return order;
+        order.setTotalPrice(totalPrice);
+        return orderRepository.save(order);
     }
 
     @Transactional(rollbackFor = {Exception.class})
